@@ -36,14 +36,24 @@ dependencies:
   intl: ^0.18.1
 
   # Google
-  google_maps_flutter: ^2.2.3
+  google_maps_flutter: ^2.10.0
+  geolocator: ^13.0.2
+  vietmap_flutter_gl: ^3.1.1
+  vietmap_flutter_plugin: ^0.3.0
+  location: ^6.0.2
+  flutter_polyline_points: ^2.1.0
+
   google_fonts: 5.1.0
 
   # Firebase
   firebase_core: ^3.4.1
   firebase_auth: ^5.2.1
-  firebase_messaging: ^15.1.1
+  firebase_messaging: ^15.1.6
   firebase_crashlytics: ^4.1.1
+  cloud_firestore: ^5.5.1
+
+  # Notification
+  flutter_local_notifications: ^18.0.1
 
   # CAMERA
   camera: ^0.10.5+9
@@ -54,6 +64,10 @@ dependencies:
   # ICON
   cupertino_icons: ^1.0.6
   iconly: ^1.0.1
+
+  #DATE PICKER
+  time_picker_spinner_pop_up: ^2.0.0
+  date_picker_timeline: ^1.2.6
   
   # UI 
   flutter_cache_manager: ^3.3.1
@@ -79,6 +93,14 @@ dependencies:
   animated_snack_bar: ^0.4.0
   image: ^4.3.0
   scrollable_positioned_list: ^0.3.8
+  dotted_line: ^3.2.3
+  flutter_confetti: ^0.3.4
+  loading_animation_widget: ^1.3.0
+  flutter_dotenv: ^5.2.1
+  easy_stepper: ^0.8.5+1
+  
+  
+ 
 
 dependency_overrides:
   flutter_svg: 1.1.6
@@ -92,7 +114,7 @@ dev_dependencies:
     git:
       url: https://github.com/imclerran/flutter-icons.git
       ref: dart-3
-
+  flutter_launcher_icons: ^0.14.2
   flutter_lints: ^3.0.0
 
 flutter:
@@ -107,13 +129,13 @@ flutter:
   fonts:
     - family: Google Sans
       fonts:
-        - asset: assets/fonts/PloniBold.otf
+        - asset: assets/fonts/OpenSans-Bold.ttf
           weight: 700
-        - asset: assets/fonts/PloniRegular.otf
+        - asset: assets/fonts/OpenSans-Regular.ttf
           weight: 400
     - family: Google Sans Medium
       fonts:
-        - asset: assets/fonts/PloniMedium.otf
+        - asset: assets/fonts/OpenSans-Medium.ttf
           weight: 400
     - family: Oleo Script Swash Caps
       fonts:
