@@ -1,15 +1,21 @@
 import 'dart:async';
-
-import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_crashlytics/firebase_crashlytics.dart';
-import 'package:flutter/material.dart';
-import 'package:kiwis_flutter/core/constants/constants.dart';
-import 'package:kiwis_flutter/my_app.dart';
-import 'package:kiwis_flutter/services/services.dart';
-import 'package:localize_and_translate/localize_and_translate.dart';
 import 'firebase_options.dart';
 
+import 'package:kiwis_flutter/my_app.dart';
+import 'package:kiwis_flutter/services/services.dart';
+import 'package:kiwis_flutter/services/map.service.dart';
+import 'package:kiwis_flutter/core/constants/constants.dart';
+import 'package:kiwis_flutter/services/geolocator.service.dart';
+import 'package:kiwis_flutter/services/notification.service.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
+import 'package:localize_and_translate/localize_and_translate.dart';
+
 void main() async {
+  await dotenv.load(fileName: ".env");
   await runZonedGuarded(
     () async {
       WidgetsFlutterBinding.ensureInitialized();
@@ -25,6 +31,9 @@ void main() async {
       );
       await LocalStorageService.getPrefs();
 
+      await Get.putAsync(() => GeolocatorService().init());
+      await Get.putAsync(() => MapService().init());
+      await Get.putAsync(() => NotificationService().init());
       // Run app!
       runApp(
         const LocalizedApp(
