@@ -3,40 +3,66 @@ import 'dart:io';
 import 'package:animated_snack_bar/animated_snack_bar.dart';
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_confetti/flutter_confetti.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:kiwis_flutter/core/constants/app_export.dart';
 import 'package:kiwis_flutter/core/base/base.controller.dart';
+import 'package:kiwis_flutter/core/constants/constants.dart';
+import 'package:kiwis_flutter/core/manager/manager.socket.dart';
 import 'package:kiwis_flutter/models/post.model.dart';
 import 'package:kiwis_flutter/models/user.models.dart';
 import 'package:kiwis_flutter/requests/auth.request.dart';
 import 'package:kiwis_flutter/requests/user.request.dart';
 import 'package:kiwis_flutter/requests/upload_realtime.request.dart';
 import 'package:kiwis_flutter/services/services.dart';
+import 'package:kiwis_flutter/views/home/widgets/account.content.dart';
+import 'package:kiwis_flutter/views/home/widgets/change_name.content.dart';
+import 'package:kiwis_flutter/views/home/widgets/change_password.content.dart';
+import 'package:kiwis_flutter/views/home/widgets/delete_account.content.dart';
 import 'package:kiwis_flutter/views/home/widgets/menu_content.dart';
 import 'package:kiwis_flutter/views/home/widgets/friend_content.dart';
 import 'package:image/image.dart' as img;
 
 class HomeController extends BaseController {
   /// Variables
-  AuthRequest _authRequest = AuthRequest();
-  UploadRealtimeRequest _uploadRealtimeRequest = UploadRealtimeRequest();
-  UserRequest _userRequest = UserRequest();
-  Rx<UserModel> user = UserModel().obs;
+  final ImagePicker _imagePicker = ImagePicker();
+  final AuthRequest _authRequest = AuthRequest();
+  final UserRequest _userRequest = UserRequest();
+  final UploadRealtimeRequest _uploadRealtimeRequest = UploadRealtimeRequest();
 
   // Home content variables
+  late CameraController cameraController;
+  final ConfettiController confettiController = ConfettiController();
+  final TextEditingController commentTEC = TextEditingController();
   final TextEditingController captionTEC = TextEditingController();
+  final TextEditingController firstNameTEC = TextEditingController();
+  final TextEditingController lastNameTEC = TextEditingController();
+  final TextEditingController currentPasswordTEC = TextEditingController();
+  final TextEditingController newPasswordTEC = TextEditingController();
+  final TextEditingController confirmPasswordTEC = TextEditingController();
+  final TextEditingController friendIdTEC = TextEditingController();
+
+  final Rx<UserModel> user = UserModel().obs;
+  final RxBool isFlashOn = false.obs;
+  final RxBool isRotate = false.obs;
   final RxBool onPost = false.obs;
+  final RxBool isDeleteAccount = false.obs;
+  final RxBool isOnchangeAvatar = false.obs;
+  final RxBool isCameraInitialized = false.obs;
+  final RxInt selectedCameraIndex = 0.obs;
   final Rx<XFile> imageXFile = XFile("").obs;
+  final Rx<File> avatar = File('').obs;
   final RxList<dynamic> friends = RxList<dynamic>([]);
   final RxList<PostModel> posts = RxList<PostModel>([]);
-
-  // Friend content variables
-  final TextEditingController friendIdTEC = TextEditingController();
+  List<CameraDescription> cameras = [];
+  List<String> emojis = ['😊', '😂', '❤️', '😍', '🤔', '🔥'];
 
   /// Handle
   @override
   void onInit() async {
     super.onInit();
     showLoading();
+    listenPost();
     await initializeCamera();
     await getPosts();
     user.value = AuthServices.currentUser!;
@@ -45,6 +71,14 @@ class HomeController extends BaseController {
 
   @override
   void onClose() {
+    commentTEC.dispose();
+    captionTEC.dispose();
+    firstNameTEC.dispose();
+    lastNameTEC.dispose();
+    currentPasswordTEC.dispose();
+    newPasswordTEC.dispose();
+    confirmPasswordTEC.dispose();
+    friendIdTEC.dispose();
     if (cameraController.value.isInitialized) {
       cameraController.dispose();
     }
@@ -52,8 +86,15 @@ class HomeController extends BaseController {
   }
 
   /// Home content
+  void listenPost() {
+    ManagerSocket.socket?.on(AppAPI.socketReceivePost, (data) {
+      final PostModel post = PostModel.fromJson(data);
+      posts.value.insert(0, post);
+      posts.refresh();
+    });
+  }
 
-  Future<void> onPressedPost(BuildContext context) async {
+  Future<void> handlePost(BuildContext context) async {
     try {
       if (imageXFile.value.path.isNotEmpty) {
         final response = await _uploadRealtimeRequest.uploadRealtimeRequest(
@@ -63,6 +104,14 @@ class HomeController extends BaseController {
         if (response.allGood) {
           captionTEC.clear();
           onPost.value = false;
+
+          final PostModel post = PostModel.fromJson(response.body);
+          posts.value.insert(0, post);
+          posts.refresh();
+
+          ManagerSocket.sendPost(
+            postId: post.realtimePostId!,
+          );
         } else {
           AnimatedSnackBar.material(
             response.error!,
@@ -96,12 +145,15 @@ class HomeController extends BaseController {
     );
   }
 
+  void onPressedAccount(BuildContext context) {
+    showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      builder: (context) => AccountContent(),
+    );
+  }
+
   /// Camera
-  late CameraController cameraController;
-  var isCameraInitialized = false.obs;
-  RxBool isFlashOn = false.obs;
-  RxInt selectedCameraIndex = 0.obs;
-  List<CameraDescription> cameras = [];
 
   Future<void> initializeCamera() async {
     try {
@@ -112,7 +164,10 @@ class HomeController extends BaseController {
       }
 
       cameraController = CameraController(
-          cameras[selectedCameraIndex.value], ResolutionPreset.high);
+        cameras[selectedCameraIndex.value],
+        ResolutionPreset.high,
+        enableAudio: false,
+      );
 
       await cameraController.initialize();
       isCameraInitialized.value = true;
@@ -122,26 +177,23 @@ class HomeController extends BaseController {
     }
   }
 
-  void closeOnPost() {
+  void closeOnPost() async {
     onPost.value = false;
+    isFlashOn.value = false;
+    await cameraController.setFlashMode(
+      FlashMode.off,
+    );
     imageXFile.value = XFile("");
     captionTEC.clear();
   }
 
   void toggleFlash() async {
-    try {
-      isFlashOn.value = !isFlashOn.value;
-      await cameraController.setFlashMode(
-        isFlashOn.value ? FlashMode.torch : FlashMode.off,
-      );
-    } catch (e) {
-      print("Error toggling flash: $e");
-      isFlashOn.value = !isFlashOn.value; // Revert flash state if toggle fails
-    }
+    isFlashOn.value = !isFlashOn.value;
   }
 
   Future<void> toggleRotate() async {
     try {
+      isRotate.value = !isRotate.value;
       isCameraInitialized.value = false;
       selectedCameraIndex.value =
           (selectedCameraIndex.value + 1) % cameras.length;
@@ -162,10 +214,35 @@ class HomeController extends BaseController {
     }
   }
 
+  Future<void> sendComment(String postId) async {
+    if (commentTEC.text.isNotEmpty) {
+      ManagerSocket.sendComment(
+        senderId: user.value.userId!,
+        postId: postId,
+        messageText: commentTEC.text,
+      );
+      commentTEC.clear();
+    }
+  }
+
+  Future<void> sendEmoji(BuildContext context, String emoji) async {
+    Confetti.launch(
+      context,
+      options: ConfettiOptions(particleCount: 20),
+      particleBuilder: (index) => Emoji(emoji: emoji),
+    );
+  }
+
   Future<void> takePicture() async {
-    if (!isCameraInitialized.value) return null;
+    if (!isCameraInitialized.value) return;
 
     try {
+      // Bật flash trước khi chụp
+
+      if (isFlashOn.value) {
+        await cameraController.setFlashMode(FlashMode.torch);
+      }
+
       // Chụp ảnh và lưu vào file
       final XFile originalImage = await cameraController.takePicture();
       final File originalFile = File(originalImage.path);
@@ -178,11 +255,14 @@ class HomeController extends BaseController {
       }
 
       // Lật ảnh theo chiều ngang
-      final img.Image flippedImage = img.flipHorizontal(image);
+      img.Image? flippedImage;
+      if (isRotate.value) {
+        flippedImage = img.flipHorizontal(image);
+      }
 
       // Transform kích thước ảnh (ví dụ: 200x200)
       final img.Image resizedImage = img.copyResize(
-        flippedImage,
+        flippedImage ?? image,
         width: 200,
         height: 200,
       );
@@ -196,9 +276,12 @@ class HomeController extends BaseController {
       // Cập nhật giá trị imageXFile
       imageXFile.value = XFile(resizedFile.path);
       onPost.value = true;
+
+      // Tắt flash sau khi chụp
+      await cameraController.setFlashMode(FlashMode.off);
     } catch (e) {
       print("Error taking picture: $e");
-      return null;
+      return;
     }
   }
 
@@ -218,12 +301,9 @@ class HomeController extends BaseController {
 
   Future<void> onPressedLogout(BuildContext context) async {
     try {
-      final reponse = await _authRequest.revokeFirebaseTokenRequest();
-      if (reponse.allGood) {
-        await _clearLocalStorage();
-
-        Get.offAndToNamed(Routes.SIGN_IN);
-      }
+      await _authRequest.revokeFirebaseTokenRequest();
+      await _clearLocalStorage();
+      Get.offAndToNamed(Routes.SIGN_IN);
     } catch (err) {
       print(err);
       Get.offAndToNamed(Routes.SIGN_IN);
@@ -281,5 +361,193 @@ class HomeController extends BaseController {
         type: AnimatedSnackBarType.warning,
       ).show(context);
     }
+  }
+
+  // Account setting content
+  void handleDisplayName(BuildContext context) async {
+    if (firstNameTEC.text.isEmpty || firstNameTEC.text == '') {
+      AnimatedSnackBar.material(
+        'First name cannot be empty'.tr,
+        type: AnimatedSnackBarType.warning,
+      ).show(context);
+    } else {
+      final response = await _userRequest.updateDisplayName(
+        firstNameTEC.text,
+        lastNameTEC.text,
+      );
+      if (response.allGood) {
+        user.value.firstName = firstNameTEC.text;
+        user.value.lastName = lastNameTEC.text;
+        await AuthServices.saveUser(user.value.toJson());
+
+        firstNameTEC.clear();
+        lastNameTEC.clear();
+        Get.back();
+
+        Get.snackbar(
+          'Success'.tr,
+          'Change name success'.tr,
+          colorText: appTheme.gray50,
+        );
+      }
+    }
+  }
+
+  Future<void> handlePassword(BuildContext context) async {
+    if (currentPasswordTEC.text.isEmpty ||
+        currentPasswordTEC.text == '' ||
+        newPasswordTEC.text.isEmpty ||
+        newPasswordTEC.text == '' ||
+        confirmPasswordTEC.text.isEmpty ||
+        confirmPasswordTEC.text == '') {
+      AnimatedSnackBar.material(
+        'Password cannot be empty'.tr,
+        type: AnimatedSnackBarType.warning,
+      ).show(context);
+    } else {
+      if (newPasswordTEC.text == currentPasswordTEC.text) {
+        AnimatedSnackBar.material(
+          'New password cannot be the same as the current password'.tr,
+          type: AnimatedSnackBarType.warning,
+        ).show(context);
+      } else if (newPasswordTEC.text != confirmPasswordTEC.text) {
+        AnimatedSnackBar.material(
+          'New password and confirm password do not match'.tr,
+          type: AnimatedSnackBarType.warning,
+        ).show(context);
+      } else {
+        final response = await _authRequest.updatePasswordRequest(
+          oldPassword: currentPasswordTEC.text,
+          newPassword: newPasswordTEC.text,
+          confirmPassword: confirmPasswordTEC.text,
+        );
+        if (response.allGood) {
+          currentPasswordTEC.clear();
+          newPasswordTEC.clear();
+          confirmPasswordTEC.clear();
+          Get.back();
+          Get.snackbar(
+            'Success'.tr,
+            'Change password success'.tr,
+            colorText: appTheme.gray50,
+          );
+        } else {
+          Get.snackbar(
+            'Error'.tr,
+            response.error!,
+            colorText: appTheme.gray50,
+          );
+        }
+      }
+    }
+  }
+
+  Future<void> handleDeleteAccount(BuildContext context) async {
+    try {
+      if (isDeleteAccount.value) {
+        final response = await _authRequest.deleteAccountRequest();
+        if (response.allGood) {
+          await _authRequest.revokeFirebaseTokenRequest();
+          await _clearLocalStorage();
+          Get.offAndToNamed(Routes.SIGN_IN);
+          Get.snackbar(
+            'Success'.tr,
+            'Delete account success'.tr,
+            colorText: appTheme.gray50,
+          );
+        }
+      } else {
+        AnimatedSnackBar.material(
+          'You must agree to continue'.tr,
+          type: AnimatedSnackBarType.warning,
+        ).show(context);
+      }
+    } catch (err) {
+      print(err);
+      Get.toNamed(Routes.SIGN_IN);
+    }
+  }
+
+  Future<void> handleChangeAvatar(BuildContext context) async {
+    isOnchangeAvatar.value = false;
+    try {
+      if (avatar.value.path.isNotEmpty) {
+        final response = await _userRequest.changeAvatarRequest(
+          avatar.value.readAsBytesSync(),
+        );
+        if (response.allGood) {
+          UserModel responseUser = UserModel.fromJson(response.body);
+          user.value.avatar = responseUser.avatar;
+
+          await AuthServices.saveUser(user.value.toJson());
+          user.refresh();
+          Get.snackbar(
+            'Success'.tr,
+            'Change avatar success'.tr,
+            colorText: appTheme.gray50,
+          );
+        } else {
+          Get.snackbar(
+            'Error'.tr,
+            response.error!,
+            colorText: appTheme.gray50,
+          );
+        }
+      }
+    } catch (err) {
+      print(err);
+    }
+  }
+
+  /// Chat room methods
+  Future<void> pickAvatar() async {
+    final image = await _imagePicker.pickImage(source: ImageSource.gallery);
+    if (image != null) {
+      avatar.value = File(image.path);
+      isOnchangeAvatar.value = true;
+    }
+  }
+
+  void onPressedDeleteAccountContent(BuildContext context) {
+    showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      builder: (context) => Container(
+        padding: EdgeInsets.only(
+          bottom: MediaQuery.of(context).viewInsets.bottom,
+        ),
+        child: DeleteAccountContent(),
+      ),
+    );
+  }
+
+  void onPressedChangeNameContent(BuildContext context) {
+    showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      builder: (context) => Container(
+        padding: EdgeInsets.only(
+          bottom: MediaQuery.of(context).viewInsets.bottom,
+        ),
+        child: ChangeNameContent(),
+      ),
+    );
+  }
+
+  void onPressedPasswordContent(BuildContext context) {
+    showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      builder: (context) => Container(
+        padding: EdgeInsets.only(
+          bottom: MediaQuery.of(context).viewInsets.bottom,
+        ),
+        child: ChangePasswordContent(),
+      ),
+    );
+  }
+
+  void onPressedMyPlan() {
+    Get.toNamed(Routes.PLAN);
   }
 }
