@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:iconly/iconly.dart';
 import 'package:kiwis_flutter/core/base/base.view.dart';
 import 'package:kiwis_flutter/core/constants/app_export.dart';
 import 'package:kiwis_flutter/views/home/home_controller.dart';
@@ -15,11 +14,10 @@ class FriendContent extends BaseView<HomeController> {
   /// Section Widget
   PreferredSizeWidget appBar(BuildContext context) {
     return PreferredSize(
-      preferredSize: Size.fromHeight(Get.height * 0.13), // Set desired height
+      preferredSize: Size.fromHeight(Get.height * 0.13),
       child: Container(
         padding: EdgeInsets.only(top: 23),
-        margin: EdgeInsets.symmetric(
-            horizontal: 16.0, vertical: 16), // Adjust padding as needed
+        margin: EdgeInsets.symmetric(horizontal: 16.0, vertical: 16),
         child: CustomAppBar(
           leadingWidth: 44.h,
           leading: AppbarLeadingIconbutton(
@@ -148,6 +146,7 @@ class FriendContent extends BaseView<HomeController> {
                     fit: BoxFit.cover,
                     height: 47.h,
                     width: 47.h,
+                    radius: BorderRadiusStyle.roundedBorder24,
                   ),
                 ],
               ),
