@@ -19,6 +19,9 @@ import 'package:kiwis_flutter/views/message/widgets/create_group.content.dart';
 import 'package:kiwis_flutter/views/message/widgets/group_name_content.dart';
 import 'package:kiwis_flutter/views/message/widgets/members.content.dart';
 import 'package:kiwis_flutter/views/message/widgets/setting_chat_room_content.dart';
+import 'package:kiwis_flutter/views/message/widgets/sharemap.content.dart';
+import 'package:kiwis_flutter/views/plan/plan_view.dart';
+import 'package:kiwis_flutter/views/plan/widgets/plan_create.content.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
 
 class MessageController extends BaseController {
@@ -34,14 +37,12 @@ class MessageController extends BaseController {
   final TextEditingController editGroupNameTEC = TextEditingController();
 
   // Variables
-  // Message content
-  Rx<UserModel> user = UserModel().obs;
-  RxList<GroupModel> groups = <GroupModel>[].obs;
-  // Chat room content
-  RxInt selectedIndex = 0.obs;
-  Rx<MessageModel> currentMessage = MessageModel().obs;
+  RxInt currentGroupIndex = 0.obs;
   RxBool isOnchangeAvatar = false.obs;
+  Rx<UserModel> user = UserModel().obs;
   Rx<File> editGroupAvatar = File('').obs;
+  RxList<GroupModel> groups = <GroupModel>[].obs;
+  Rx<MessageModel> currentMessage = MessageModel().obs;
 
   @override
   void onInit() {
@@ -56,9 +57,11 @@ class MessageController extends BaseController {
     super.onClose();
   }
 
+  GroupModel get currentGroup => groups.value[currentGroupIndex.value];
+
   /// Message methods
   void onPressedChanel(BuildContext context, int index) {
-    selectedIndex.value = index;
+    currentGroupIndex.value = index;
     showModalBottomSheet(
       context: context,
       isScrollControlled: true,
@@ -79,7 +82,7 @@ class MessageController extends BaseController {
     ManagerSocket.socket?.on(AppAPI.socketReceiveGroupMessage, (data) {
       print('Received group message: $data');
       final message = MessageModel.fromJson(data);
-      final group = groups.value[selectedIndex.value];
+      final group = groups.value[currentGroupIndex.value];
       group.messages?.add(message);
       groups.refresh();
     });
@@ -143,6 +146,8 @@ class MessageController extends BaseController {
   }
 
   /// Chat room methods
+
+  // Handle
   Future<void> pickAvatar() async {
     final image = await _imagePicker.pickImage(source: ImageSource.gallery);
     if (image != null) {
@@ -151,10 +156,39 @@ class MessageController extends BaseController {
     }
   }
 
+  Future<void> handleLeaveGroup() async {
+    try {
+      final response = await _groupRequest.leaveGroupRequest(
+        groupId: groups.value[currentGroupIndex.value].groupId!,
+      );
+      if (response.allGood) {
+        Get.offAllNamed(Routes.MAIN);
+        Get.snackbar("Leave group", "Leave group successfully");
+      }
+    } catch (err) {
+      Get.snackbar("Leave group", err.toString());
+    }
+  }
+
+  Future<void> handleBlockUser() async {
+    try {
+      final response = await _groupRequest.blockUserRequest(
+        userId:
+            groups.value[currentGroupIndex.value].members!.first.user!.userId!,
+      );
+      if (response.allGood) {
+        Get.offAllNamed(Routes.MAIN);
+        Get.snackbar("Block user", "Block user successfully");
+      }
+    } catch (err) {
+      Get.snackbar("Block user", err.toString());
+    }
+  }
+
   Future<void> handleChangeGroupName(BuildContext context) async {
     try {
       final response = await _groupRequest.editGroupRequest(
-        groupId: groups.value[selectedIndex.value].groupId!,
+        groupId: groups.value[currentGroupIndex.value].groupId!,
         name: editGroupNameTEC.text,
         file: editGroupAvatar.value.readAsBytesSync(),
       );
@@ -165,8 +199,8 @@ class MessageController extends BaseController {
         editGroupAvatar.value = File('');
         editGroupNameTEC.clear();
         // Update group
-        groups.value[selectedIndex.value].name = group.name;
-        groups.value[selectedIndex.value].avatar = group.avatar;
+        groups.value[currentGroupIndex.value].name = group.name;
+        groups.value[currentGroupIndex.value].avatar = group.avatar;
         groups.refresh();
 
         Get.back();
@@ -188,47 +222,39 @@ class MessageController extends BaseController {
     }
   }
 
-  Future<void> handleLeaveGroup(BuildContext context) async {
-    try {
-      final response = await _groupRequest.leaveGroupRequest(
-        groupId: groups.value[selectedIndex.value].groupId!,
+  void sendMessage() {
+    if (messageTEC.text.isNotEmpty) {
+      ManagerSocket.sendMessage(
+        senderId: user.value.userId!,
+        groupId: groups.value[currentGroupIndex.value].groupId!,
+        messageText: messageTEC.text,
       );
-      if (response.allGood) {
-        Get.back();
-        AnimatedSnackBar.material(
-          "Leave group successfully",
-          type: AnimatedSnackBarType.success,
-        ).show(context);
-      } else {
-        AnimatedSnackBar.material(
-          response.error!,
-          type: AnimatedSnackBarType.error,
-        ).show(context);
-      }
-    } catch (err) {
-      AnimatedSnackBar.material(
-        err.toString(),
-        type: AnimatedSnackBarType.error,
-      ).show(context);
+      messageTEC.clear();
     }
   }
 
-  void sendMessage() {
-    ManagerSocket.sendMessage(
-      user.value.userId!,
-      groups.value[selectedIndex.value].groupId!,
-      messageTEC.text,
+  void onPressedShareLocation(BuildContext context) {
+    showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      builder: (context) => ShareMapContent(),
     );
-    messageTEC.clear();
+  }
+
+  void onPressedCreatePlan() {
+    Get.toNamed(
+      Routes.PLAN,
+      arguments: groups.value[currentGroupIndex.value].groupId!,
+    );
   }
 
   String getGroupName() {
-    final group = groups.value[selectedIndex.value];
+    final group = groups.value[currentGroupIndex.value];
     return group.name ?? group.members!.first.user!.fullName;
   }
 
   String getAvatarGroup() {
-    final group = groups.value[selectedIndex.value];
+    final group = groups.value[currentGroupIndex.value];
     if (group.avatar != null) {
       return group.avatar!.imageUrl!;
     } else if (isGroupChat()) {
@@ -239,7 +265,7 @@ class MessageController extends BaseController {
   }
 
   bool isGroupChat() {
-    final group = groups.value[selectedIndex.value];
+    final group = groups.value[currentGroupIndex.value];
     return group.isGroupChat();
   }
 
