@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:iconly/iconly.dart';
 
 import 'package:kiwis_flutter/core/constants/app_export.dart';
-import 'package:kiwis_flutter/core/base/base.view.dart';
 import 'package:kiwis_flutter/views/message/widgets/chat_bubble.widget.dart';
 import 'package:kiwis_flutter/views/message/message_controller.dart';
 import 'package:kiwis_flutter/widgets/app_bar/app_bar_leadingiconbutton.dart';
@@ -12,7 +11,7 @@ import 'package:kiwis_flutter/widgets/app_bar/custom_app_bar.dart';
 import 'package:kiwis_flutter/widgets/custom_text_form_field.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
 
-class ChatRoomContent extends BaseView<MessageController> {
+class ChatRoomContent extends GetView<MessageController> {
   ChatRoomContent({super.key});
 
   Widget? floatingActionButton(BuildContext context) {
@@ -23,20 +22,26 @@ class ChatRoomContent extends BaseView<MessageController> {
         mainAxisAlignment: MainAxisAlignment.center,
         mainAxisSize: MainAxisSize.min,
         children: [
-          Container(
-            width: 32.h,
-            height: 32.h,
-            child: Icon(
-              IconlyBold.location,
-              color: appTheme.blue300,
+          GestureDetector(
+            onTap: () => controller.onPressedShareLocation(context),
+            child: Container(
+              width: 32.h,
+              height: 32.h,
+              child: Icon(
+                IconlyBold.location,
+                color: appTheme.blue300,
+              ),
             ),
           ),
-          Container(
-            width: 32.h,
-            height: 32.h,
-            child: Icon(
-              IconlyBold.camera,
-              color: appTheme.blue300,
+          GestureDetector(
+            onTap: () => controller.onPressedShareLocation(context),
+            child: Container(
+              width: 32.h,
+              height: 32.h,
+              child: Icon(
+                IconlyBold.camera,
+                color: appTheme.blue300,
+              ),
             ),
           ),
           SizedBox(width: 8.h),
@@ -94,12 +99,12 @@ class ChatRoomContent extends BaseView<MessageController> {
             ),
           ),
           actions: [
-            AppbarTrailingIconbutton(
+            AppbarLeadingIconbutton(
               onTap: () => controller.showModalCalenderSheet(context),
               imagePath: ImageConstant.svgCalander,
               margin: EdgeInsets.only(right: 8.h),
             ),
-            AppbarTrailingIconbutton(
+            AppbarLeadingIconbutton(
               onTap: () => controller.showModalSettingChatRoomSheet(context),
               imagePath: ImageConstant.svgMore,
               margin: EdgeInsets.only(right: 8.h),
@@ -111,63 +116,108 @@ class ChatRoomContent extends BaseView<MessageController> {
   }
 
   @override
-  Widget body(BuildContext context) {
-    return Container(
-      height: Get.height,
-      width: double.maxFinite,
-      padding: EdgeInsets.symmetric(horizontal: 16.h),
-      child: Stack(
-        children: [
-          Align(
-            alignment: Alignment.topCenter,
-            child: SizedBox(
-              height: Get.height * 0.78,
-              child: Obx(() {
-                return ScrollablePositionedList.builder(
-                  itemScrollController: controller.scrollController,
-                  initialScrollIndex: controller.groups
-                      .value[controller.selectedIndex.value].messages!.length,
-                  padding: const EdgeInsets.all(8.0),
-                  itemCount: controller.groups
-                      .value[controller.selectedIndex.value].messages!.length,
-                  itemBuilder: (context, index) {
-                    final messages = controller
-                        .groups.value[controller.selectedIndex.value].messages!;
-                    final currentMessage = messages[index];
-                    final isGroupStart = index == 0 ||
-                        messages[index - 1].senderId != currentMessage.senderId;
-                    final isGroupEnd = index == messages.length - 1 ||
-                        messages[index + 1].senderId != currentMessage.senderId;
+  Widget build(BuildContext context) {
+    return Scaffold(
+      floatingActionButton: floatingActionButton(context),
+      appBar: appBar(context),
+      body: SizedBox(
+        width: double.maxFinite,
+        child: Container(
+          height: Get.height,
+          child: Stack(
+            alignment: Alignment.center,
+            children: [
+              CustomImageView(
+                imagePath: ImageConstant.imgEllipse2005,
+                height: 280.h,
+                width: 274.h,
+                alignment: Alignment.topRight,
+              ),
+              CustomImageView(
+                imagePath: ImageConstant.imgEllipse2006Green600516x342,
+                height: 340.h,
+                width: 240.h,
+                alignment: Alignment.bottomLeft,
+              ),
+              SafeArea(
+                child: SingleChildScrollView(
+                  child: Container(
+                    height: Get.height,
+                    width: double.maxFinite,
+                    padding: EdgeInsets.symmetric(horizontal: 16.h),
+                    child: Stack(
+                      children: [
+                        Align(
+                          alignment: Alignment.topCenter,
+                          child: SizedBox(
+                            height: Get.height * 0.78,
+                            child: Obx(() {
+                              return ScrollablePositionedList.builder(
+                                itemScrollController:
+                                    controller.scrollController,
+                                initialScrollIndex: controller
+                                    .groups
+                                    .value[controller.currentGroupIndex.value]
+                                    .messages!
+                                    .length,
+                                padding: const EdgeInsets.all(8.0),
+                                itemCount: controller
+                                    .groups
+                                    .value[controller.currentGroupIndex.value]
+                                    .messages!
+                                    .length,
+                                itemBuilder: (context, index) {
+                                  final messages = controller
+                                      .groups
+                                      .value[controller.currentGroupIndex.value]
+                                      .messages!;
+                                  final currentMessage = messages[index];
+                                  final isGroupStart = index == 0 ||
+                                      messages[index - 1].senderId !=
+                                          currentMessage.senderId;
+                                  final isGroupEnd =
+                                      index == messages.length - 1 ||
+                                          messages[index + 1].senderId !=
+                                              currentMessage.senderId;
 
-                    return ChatBubble(
-                      message: currentMessage,
-                      showAvatar: isGroupStart,
-                      isGroup: controller.isGroupChat(),
-                      borderRadius: BorderRadius.only(
-                        topLeft: Radius.circular((isGroupStart ? 8 : 8)),
-                        topRight: Radius.circular(
-                          !currentMessage.isMe()
-                              ? (isGroupStart ? 8 : 8)
-                              : (isGroupStart ? 8 : 4),
+                                  return ChatBubble(
+                                    message: currentMessage,
+                                    showAvatar: isGroupStart ||
+                                        currentMessage.isComment(),
+                                    isGroup: controller.isGroupChat(),
+                                    borderRadius: BorderRadius.only(
+                                      topLeft: Radius.circular(
+                                          (isGroupStart ? 8 : 8)),
+                                      topRight: Radius.circular(
+                                        !currentMessage.isMe()
+                                            ? (isGroupStart ? 8 : 8)
+                                            : (isGroupStart ? 8 : 4),
+                                      ),
+                                      bottomLeft: Radius.circular(
+                                        !currentMessage.isMe()
+                                            ? (isGroupEnd ? 8 : 4)
+                                            : (isGroupEnd ? 4 : 8),
+                                      ),
+                                      bottomRight: Radius.circular(
+                                        !currentMessage.isMe()
+                                            ? (isGroupEnd ? 4 : 8)
+                                            : (isGroupEnd ? 8 : 4),
+                                      ),
+                                    ),
+                                  );
+                                },
+                              );
+                            }),
+                          ),
                         ),
-                        bottomLeft: Radius.circular(
-                          !currentMessage.isMe()
-                              ? (isGroupEnd ? 8 : 4)
-                              : (isGroupEnd ? 4 : 8),
-                        ),
-                        bottomRight: Radius.circular(
-                          !currentMessage.isMe()
-                              ? (isGroupEnd ? 4 : 8)
-                              : (isGroupEnd ? 8 : 4),
-                        ),
-                      ),
-                    );
-                  },
-                );
-              }),
-            ),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ],
           ),
-        ],
+        ),
       ),
     );
   }
